requirements.txt
together-ai
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-openai>=0.2,<0.3
langchain-groq>=0.2,<0.3
langchain-text-splitters>=0.3,<0.4
streamlit~=1.32.2
dspy-ai>=2.4.0
groq>=0.4.1
openai>=1.0.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
huggingface-hub>=0.16.0
chromadb>=0.4.15
pypdf>=3.15.0
python-docx>=0.8.11
openpyxl>=3.1.2
pandas>=2.0.0
nltk>=3.8.1
spacy>=3.6.0
tiktoken>=0.5.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.15.0
pyyaml~=6.0.2
python-dotenv~=1.1.0
loguru>=0.7.0
pydantic>=2.0.0
marshmallow>=3.20.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0